Add Listing and Newspaper
These are shaping up to be business entities. I'm leaving them in
app/models because that is where people will typically look for this
kind of thing in a Rails app, but I am tempted to put them in
app/entities since I plan to separate the entity/presenter/validator
responsibilities.