Revert to typical Rails approaches for examples
This commit makes a few changes to take the master branch to a more
typical Rails approach, with models directly inheriting from
ApplicationRecord and traditional controller/view implementation. The
idea is that the policy examples will be more approachable if they work
with a more traditional layout.

There are also many things to figure out with the repository and use
case approaches, so it is partially expedient to start traditional and
partially about setting up a refactoring/evolution exercise.

This also adds Slim for views and an .envrc for direnv use.