AEIM-725 - Condition delete link
This backtracks on rendering only with locals (for now) to make it
easier to deliver the policy to the view. There is plenty of room for
adding convenience methods and view models / presenters, but the way it
was shaped so far was pretty clumsy.

This also addresses ActiveRecord behavior with a non-User null object.
The POROs are deleted and there is a `User.nobody` method for now. There
are some other options, including using something like the active_null
gem, forcing not-null on the listing, doing the coalescing at a
presenter, and so on. Like most of this commit, this is not a firm
implementation decision, but an expedient one for spiking.