Refactor GrantResolver and its tests with inversion
The dummy grant data was held in the grants method, which now delegates to
a grant repository, which can be injected for testing. That repository
will likely be a wrapper for an ActiveRecord model, but we aren't yet
sure when or where that dependency should actually surface.

Also, the grant resolver tests are refactored to mock and inject the
resolvers. The name "grant resolver" is maybe not quite right, but I am
happy to keep it for now.