Clean up SubjectResolver to use UserDirectory
 * Refactor to extract dummy data from SubjectResolver
 * Change SubjectResolver to not bind to a user
 * Add empty implementation of UserDirectory
 * Inject dummy data with a fake directory

The reason to not bind a SubjectResolver to a given user is for eventual
API consistency. I expect that there will be multiple values that must
be resolved and binding the resolver to a given value (user) will force
object construction details into the callsite. It might be fine to use a
factory for this and have one resolver for each value (relying on a
service registry or DI container for dependencies like the
UserDirectory), but that doesn't seem to be the right approach for now.
Retaining a resolver that can work on multiple values seems better.