Rework CredentialResolver to use permission map
Similar to the work to have SubjectResolver delegate attribute lookup to
a user directory, this commit has the CredentialResolver delegate to a
PermissionMapper. This serves two purposes:

 1. It gives a way to inject the dummy data into the CredentialResolver.
 2. It makes room for implementing responsibilities of managing how actions
    translate to permissions and which roles would grant them outside of
    the generic resolution behavior.

In some applications, it may be entirely sufficient to use a passthrough
implementation that simply prepends "permission" before the action names
and does nothing with roles. In some, it may be sufficient to implement
a simple role-permission mapping in a class or config file. This design
affords object-oriented extension around the core
Subject-Credential-Resource model.

For one specific example, a particularly RESTful application may benefit
from mapping all CRUD actions to a `manage` permission. This is a little
different than a role, implying that anyone who is granted the
permission may take a range of actions. This could help simplify the
policies and UI, where a user or role could be granted `manage` for a
type or entity. This approach affords hierarchical/inherited permissions
without dictating multiple-role, role-per-object, or role-per-type
semantics. For example, a single-valued system role of `editor` may
grant `manage` permission on multiple types, transitively granting
`show`, `create`, `update`, and `destroy`.