Refactor ResourceResolver to not capture target
To keep the API consistent across the resolvers, and at the risk of a
data clump on the helper methods that handle the object to be resolved,
this adds a parameter to `ResourceResolver#resolve` and takes it off of
the constructor.

We may want to reverse this decision once there is some infrastructure
for wiring up the default collaborators for the resolvers -- at which
point the call-site complexity of creating resolvers would be low enough
to justify making the object captive. It's possible that the service
locator / factory work would have been a shorter path than removing the
captive values, but this keeps the API simple and very amenable to
caching for now.