Add a couple of presenters and policies to facilitate view checks
Particularly for lists, having each object bound to a policy for the
instance is very handy for checking access control in views. Presenters
provide a handy way to couple any view logic with access to the policy,
rather than using a helper method from the view that would have to look
up or create the policy objects.

There are a few implementation examples here:

1. A presenter method that renders a link or nothing
2. A presenter method that render a link or text
3. Template logic that retrieves the policy from the presenter for a
   rendering decision there

In general, the logic-less approach to templates (1 and 2 above) is a
little easier to test and keep consistent across an app by putting the
conditionals in ruby files. An example could be how a username might be
rendered with a link for any other logged in user, but as plain text for
guests. Just calling the presenter method that renders it the right way
is consistent and explanatory as opposed to wrapping each use in a
conditional or helper call.

However, if there is significant markup that needs to be conditioned for
a section, migrating it to a presenter poses an inverse challenge. In
those cases, a decent middle-ground is to keep the logic of the
decisions in named methods on the presenter, so they can be tested, with
the complex markup for each alternative in the template. This approach
puts the behavioral selection in readily tested code without coupling it
to the exact contents to be rendered.

This commit also introduces a CollectionPolicy as a partner for the
ResourcePolicy. This serves as the base class for policies that deal
with types or collections of resources. To compare with the Pundit
model, this also serves as the place for implementing the "policy
scope"; that is, the code to select the appropriate filters for items of
that type for the user. Another difference is that the type-based access
decisions are moved here (index and new) as opposed to residing on the
combination instance/class policy.