Normalize and add convenience to presenters
This still needs tests and comments for the basis of usage and best
practices documentation, but I am quite happy with this design.

The impetus for the design change was that we want resources to be able
to have a default presenter. Specifically, when displaying a listing,
the owner (User) should be presented so any "helpers" (custom presenter
methods, not global helpers) are available consistently. However, we
don't want the listing controller, presenter, or policy to need to know
which presenter is the "base" for a User object or which policy should
be used by default. To achieve that, we need to invoke a collaborator
(either the object itself or a third party). It would be simpler to put
the identity knowledge on the model class itself, but that is extended
responsibility whereas it now knows nothing about how it is presented or
authorized. The factory holds the mapping and is available to provide
the base/default presenter, including an appropriate policy.

There is a lot of work in this commit:

- Add ResourcePresenter and CollectionPresenter to mirror the policy
  base classes and provide convenience
- Take only the policy (not resource or collection) in these presenters
  rather than pass potentially disconnected objects
- Convert Listing presenters to extend the base classes
- Add UserPresenter to eliminate global display_name_or_link helper
- Add service registry with canister
- Add a presenter factory to build default presenters/policies based
  solely on the object/collection type
- Add `present` method to ApplictionController as a convenience when
  wanting to simply use the defaults
- Add NullPresenter and ReadOnlyPolicy as the super-defaults (to test
  out polymorphism/null objects here, rather than simply returning the
  object when no presenter is configured)
- Include a commented example of overriding just the `present` method on
  a collection presenter for the case where it wants to use a
  specialized item/resource presenter
- Present the associated `owner` object in the ListingPresenter, rather
  than exposing it directly (the impetus for this entire design)