Extract Vizier and use everything via gems
This adds the keycard, checkpoint, and vizier gems to the Gemfile and
extracts Vizier from this application. There are instructions in the
README for using these gems from a local working copy.

I tried a number of approaches to support auto-reloading but there were
different glitches in every combination. It's not clear how to get it
working at this point. Some things I tried were:

- Using unloadable from ActiveSupport::Dependencies
- Deleting all of the gems' files from $" and requiring the gems
- Using require_dependency for the gems
- Using load for all of the gems' files