Use a biggerer hammer
Running checkpoint:migrate and then db:migrate puts the updated schema
into schema.rb. This seems like maybe the right way to do it in general,
so I'm going to go with this for now. Travis should be able to handle
this and not need to do any migrations -- just a db:setup.