Sync up Keycard and Checkpoint schemas
This commit brings in the schema information for Keycard as well as
updating the table (in schema.rb) that the schema information for
Checkpoint is stored in. Also gitignore keycard.log.

As an aside, it updates all three gem versions. Vizier and Checkpoint
have been released, so their versions/hashes are updated here.