Support loading keycard/checkpoint database settings
This pattern puts responsibility on the app to decide whether and where
there are settings for Keycard or Checkpointto apply. Given that we're
using Ettin for settings already, this seems like the right approach.
Putting specific config file conventions in Checkpoint or Keycard
directly does not seem wise, at least as long as there is minimal
configuration to do.